Removed circular dependencie in uvue loader
